requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dataclasses-json==0.5.8
datasets==2.12.0
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
jieba==0.42.1
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.209
langchainplus-sdk==0.0.17
loguru==0.7.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
pdfminer.six==20221105
pefile==2023.2.7
Pillow==9.5.0
pinecone-client==2.2.2
protobuf==4.23.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
Pympler==1.0.1
pypdf==3.11.0
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rich==13.4.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.16
streamlit==1.23.1
streamlit-chat==0.0.2.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
Wand==0.6.11
watchdog==3.0.0
win32-setctime==1.1.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
